Change Icon to Material 3
